+++ django-api/api/authentication/models/__init__.py
@@ -1,0 +1,1 @@
+from .active_session import ActiveSession