+++ django-api/api/authentication/viewsets/__init__.py
@@ -1,0 +1,4 @@
+from .register import RegisterViewSet
+from .login import LoginViewSet
+from .active_session import ActiveSessionViewSet
+from .logout import LogoutViewSet