+++ django-api/api/user/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = "api.user.apps.UserConfig"