+++ django-api/core/urls.py
@@ -1,0 +1,5 @@
+from django.urls import path, include
+
+urlpatterns = [
+    path("api/users/", include(("api.routers", "api"), namespace="api")),
+]