+++ react-ui/src/config/constant.js
@@ -1,0 +1,1 @@
+export const API_SERVER = "http://localhost:5000/api/";