+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log
 
+## [1.0.1] 2022-01-28
+### Improvements
+
+- React UI: [React Soft UI Dashboard](https://github.com/app-generator/react-soft-ui-dashboard) **v2.0.5**
+  - `Logout` fixes 
+- API: [Django API Server](https://github.com/app-generator/api-server-django) **v1.0.1**
+  - Update for `Django==4.0.1`
+
 ## [1.0.0] 2021-10-20
 ### Initial Release
 
-- UI: [React Soft UI Dashboard](https://github.com/app-generator/react-soft-ui-dashboard) **v2.0.4**    
-- Backend Version: [Django API Server](https://github.com/app-generator/api-server-django) **v1.0.0**
+- React UI: [React Soft UI Dashboard](https://github.com/app-generator/react-soft-ui-dashboard) **v2.0.4**    
+- API: [Django API Server](https://github.com/app-generator/api-server-django) **v1.0.0**